- removed the dependency on Trajectory from ROMol and ForceField
