+++ Code/Geometry/Wrap/rdGeometry.cpp
@@ -14,13 +14,11 @@
 namespace python = boost::python;
 void wrap_point();
 void wrap_uniformGrid();
-void wrap_trajectory();
 
 BOOST_PYTHON_MODULE(rdGeometry) {
   python::scope().attr("__doc__") =
-      "Module containing geometry objects like points, grids, trajectories, etc\n";
+      "Module containing geometry objects like points, grids, etc\n";
 
   wrap_point();
   wrap_uniformGrid();
-  wrap_trajectory();
 }
