+++ Code/GraphMol/Wrap/Mol.cpp
@@ -18,7 +18,6 @@
 // ours
 #include <RDBoost/pyint_api.h>
 #include <RDBoost/Wrap.h>
-#include <Geometry/Trajectory.h>
 #include <GraphMol/RDKitBase.h>
 #include <GraphMol/QueryOps.h>
 #include <GraphMol/MolPickler.h>
@@ -334,11 +333,13 @@ struct mol_wrapper {
               python::arg("assignId") = false),
              "Add a conformer to the molecule and return the conformer ID")
 
+#if 0
         .def("AddConformersFromTrajectory", &ROMol::addConformersFromTrajectory,
              (python::arg("self"), python::arg("traj"),
               python::arg("nConf") = -1),
              "Add conformers from a Trajectory to the molecule and return\n"
              "the number of conformations that were added")
+#endif
 
         .def("GetConformer", GetMolConformer,
              (python::arg("self"), python::arg("id") = -1),
